src/api/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asynch==0.2.4
attrs==24.2.0
blinker==1.8.2
boto3==1.35.17
botocore==1.35.17
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
clickhouse-connect==0.7.18
clickhouse-driver==0.2.9
clickhouse-sqlalchemy==0.3.2
deprecation==2.1.0
fastapi==0.112.2
Flask==3.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.8.0
greenlet==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
leb128==1.0.8
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
minio==7.2.8
multidict==6.0.5
narwhals==1.6.0
numpy==1.26.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
postgrest==0.16.11
protobuf==3.20.3
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pytest==7.4.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
realtime==2.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
storage3==0.7.7
streamlit==1.38.0
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.7
zstandard==0.23.0
zstd==1.5.5.1
prefect==3.0.1

